refactor: extract domain registrars from ExtensionContextHandler
Split the 903-line god-class ExtensionContextHandler into 8 focused
domain registrars under src/extension/registrars/:

- EditorRegistrar     — toggle source code, open with Kaoto, undo/redo
- ExecutorRegistrar   — executor setup, JBang/Java checks, trusted sources
- LifecycleRegistrar  — What's New panel, recommended extensions
- IntegrationsRegistrar — view, run, project, kubernetes, Maven commands
- DeploymentsRegistrar  — view, route stop/start/resume/suspend commands
- TestsRegistrar        — view, run, init commands
- InfrastructureRegistrar — view, start/stop/logs/copy commands
- OpenApiRegistrar      — view, import command

Also extracted shared sendCommandTrackingEvent helper to trackingEvent.ts.

ExecutorInitializer.ts now accepts an IExecutorHandler interface instead
of the concrete ExtensionContextHandler class, so ExecutorRegistrar
satisfies it structurally.

extension.ts and extensionWeb.ts rewritten as composition roots.
ExtensionContextHandler.ts deleted. Test updated to target LifecycleRegistrar.

No functional changes — method bodies are identical to the original.

refactor: rename trackingEvent.ts → TrackingEvent.ts for consistent PascalCase
