+++ src/extension/registrars/TrackingEvent.ts
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2025 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { TelemetryEvent, TelemetryService } from '@redhat-developer/vscode-redhat-telemetry';
+
+export async function sendCommandTrackingEvent(telemetryService: TelemetryService | undefined, commandId: string): Promise<void> {
+	const telemetryEvent: TelemetryEvent = {
+		type: 'track',
+		name: 'command',
+		properties: {
+			identifier: commandId,
+		},
+	};
+	if (telemetryService) {
+		await telemetryService.send(telemetryEvent);
+	}
+}
